Move feature tests to seperate folders
