+++ spec/features/restaurants/index_spec.rb
@@ -63,40 +63,4 @@ RSpec.describe "As a visitor", type: :feature do
       end
     end
   end
-
-  describe 'when I visit /categories' do
-    it "loads a page" do
-      visit categories_path()
-
-      expect(page.status_code).to eq(200)
-      expect(current_path).to eq("/categories")
-    end
-
-    it "shows all data" do
-      visit categories_path()
-
-      within("table") do
-        within("#header") do
-          expect(page).to have_content("Category")
-          expect(page).to have_content("Locations in this Category")
-          expect(page).to have_content("Total chairs in this Category")
-        end
-        within("#category_other") do
-          expect(page).to have_content("other")
-          expect(page).to have_content("2")
-          expect(page).to have_content("67")
-        end
-        within("#category_ls2_large") do
-          expect(page).to have_content("ls2 large")
-          expect(page).to have_content("4")
-          expect(page).to have_content("438")
-        end
-        within("#category_ls1_large") do
-          expect(page).to have_content("ls1 large")
-          expect(page).to have_content("1")
-          expect(page).to have_content("152")
-        end
-      end
-    end
-  end
 end
